// src/modules/auth/dto/microsoft-auth-code.dto.ts
import { ApiProperty } from '@nestjs/swagger';
import { IsNotEmpty, IsOptional, IsString } from 'class-validator';

export class MicrosoftAuthCodeDto {
  @ApiProperty({
    example: '0.AbcDefGh...',
    description: 'Authorization code returned by Microsoft after login',
  })
  @IsString()
  @IsNotEmpty()
  code!: string;

  @ApiProperty({
    example: 'http://localhost:4000/runway',
    description: 'Redirect URI used during Microsoft authorization request',
  })
  @IsString()
  @IsNotEmpty()
  redirectUri!: string;

  @ApiProperty({
    example: 'random_nonce_value',
    description: 'Nonce generated by frontend and sent to Microsoft authorize URL',
    required: false,
  })
  @IsString()
  @IsOptional()
  nonce?: string;
}